# doc/openapi.yml
openapi: 3.0.3
info:
  title: Spring OpenAI Sample API
  description: Sample API for demonstrating Spring Boot with OpenAI integration
  version: 1.0.0
  contact:
    name: API Support
    email: support@arcconnects.com

servers:
  - url: http://localhost:8080
    description: Development server

paths:
  /api/health:
    get:
      summary: Health check endpoint
      operationId: healthCheck
      tags:
        - health
      responses:
        '200':
          description: Health check successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "OK"
                  timestamp:
                    type: string
                    format: date-time

  /api/chat:
    post:
      summary: Send a chat message to OpenAI
      operationId: sendChatMessage
      tags:
        - chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response from OpenAI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========================
  # ユーザー管理API
  # ========================
  /api/users:
    get:
      summary: ユーザー一覧を取得
      description: |
        登録されているユーザーの一覧を取得します。
        ページネーション、フィルタリング、ソートに対応しています。
      operationId: getUsers
      tags:
        - user
      parameters:
        - name: page
          in: query
          description: ページ番号（0から開始）
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: size
          in: query
          description: 1ページあたりの件数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: sort
          in: query
          description: |
            ソート条件（カンマ区切りで複数指定可能）
            例: name,asc または email,desc
          required: false
          schema:
            type: string
            example: "createdAt,desc"
        - name: name
          in: query
          description: 名前による部分一致検索
          required: false
          schema:
            type: string
            example: "田中"
        - name: email
          in: query
          description: メールアドレスによる部分一致検索
          required: false
          schema:
            type: string
            example: "@example.com"
        - name: status
          in: query
          description: ユーザーステータスによるフィルタ
          required: false
          schema:
            $ref: '#/components/schemas/UserStatus'
      responses:
        '200':
          description: ユーザー一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: 新規ユーザーを作成
      description: |
        新しいユーザーを作成します。
        メールアドレスは一意である必要があります。
      operationId: createUser
      tags:
        - user
      requestBody:
        required: true
        description: 作成するユーザー情報
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: ユーザー作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: リクエストボディが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '409':
          description: メールアドレスが既に使用されている
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/{id}:
    get:
      summary: ユーザー詳細を取得
      description: |
        指定されたIDのユーザー詳細情報を取得します。
        存在しないIDが指定された場合は404エラーを返します。
      operationId: getUserById
      tags:
        - user
      parameters:
        - name: id
          in: path
          description: ユーザーID
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: ユーザー詳細取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: 指定されたユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: ユーザー情報を更新
      description: |
        指定されたIDのユーザー情報を更新します。
        部分更新には対応しておらず、全項目の指定が必要です。
      operationId: updateUser
      tags:
        - user
      parameters:
        - name: id
          in: path
          description: 更新対象のユーザーID
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      requestBody:
        required: true
        description: 更新するユーザー情報
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: ユーザー更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: リクエストボディが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: 指定されたユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: メールアドレスが既に他のユーザーに使用されている
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: ユーザーを削除
      description: |
        指定されたIDのユーザーを削除します。
        物理削除ではなく、論理削除（ステータスをINACTIVEに変更）を行います。
      operationId: deleteUser
      tags:
        - user
      parameters:
        - name: id
          in: path
          description: 削除対象のユーザーID
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '204':
          description: ユーザー削除成功（レスポンスボディなし）
        '404':
          description: 指定されたユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: The message to send to OpenAI
          example: "Hello, how are you?"
        model:
          type: string
          description: OpenAI model to use
          example: "gpt-3.5-turbo"
          default: "gpt-3.5-turbo"

    ChatResponse:
      type: object
      properties:
        id:
          type: string
          description: Response ID
          example: "chatcmpl-123"
        response:
          type: string
          description: The response from OpenAI
          example: "Hello! I'm doing well, thank you for asking."
        model:
          type: string
          description: Model used
          example: "gpt-3.5-turbo"
        usage:
          $ref: '#/components/schemas/Usage'

    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
          description: Number of tokens in the prompt
          example: 12
        completion_tokens:
          type: integer
          description: Number of tokens in the completion
          example: 15
        total_tokens:
          type: integer
          description: Total number of tokens used
          example: 27

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request format"
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: Request path that caused the error
          example: "/api/chat"

    # ========================
    # ユーザー管理関連のスキーマ
    # ========================
    
    UserStatus:
      type: string
      description: ユーザーアカウントの状態
      enum:
        - ACTIVE
        - INACTIVE
        - SUSPENDED
        - PENDING
      example: ACTIVE
    
    UserRole:
      type: string
      description: ユーザーの権限レベル
      enum:
        - ADMIN
        - USER
        - GUEST
      example: USER

    CreateUserRequest:
      type: object
      description: ユーザー作成時のリクエストボディ
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: ユーザーの氏名
          minLength: 1
          maxLength: 100
          example: "田中太郎"
        email:
          type: string
          description: ユーザーのメールアドレス（一意である必要がある）
          format: email
          maxLength: 255
          example: "tanaka@example.com"
        phone:
          type: string
          description: 電話番号（ハイフンなしの半角数字）
          pattern: '^[0-9]{10,11}$'
          example: "09012345678"
        role:
          $ref: '#/components/schemas/UserRole'
        address:
          type: object
          description: 住所情報
          properties:
            postalCode:
              type: string
              description: 郵便番号（ハイフンなし）
              pattern: '^[0-9]{7}$'
              example: "1234567"
            prefecture:
              type: string
              description: 都道府県
              maxLength: 10
              example: "東京都"
            city:
              type: string
              description: 市区町村
              maxLength: 50
              example: "新宿区"
            addressLine:
              type: string
              description: 番地・建物名
              maxLength: 100
              example: "新宿1-1-1 新宿ビル101"
        birthDate:
          type: string
          description: 生年月日
          format: date
          example: "1990-04-15"
        department:
          type: string
          description: 所属部署
          maxLength: 50
          example: "開発部"

    UpdateUserRequest:
      type: object
      description: ユーザー更新時のリクエストボディ
      required:
        - name
        - email
        - status
      properties:
        name:
          type: string
          description: ユーザーの氏名
          minLength: 1
          maxLength: 100
          example: "田中次郎"
        email:
          type: string
          description: ユーザーのメールアドレス
          format: email
          maxLength: 255
          example: "tanaka.jiro@example.com"
        phone:
          type: string
          description: 電話番号（ハイフンなしの半角数字）
          pattern: '^[0-9]{10,11}$'
          example: "09087654321"
        status:
          $ref: '#/components/schemas/UserStatus'
        role:
          $ref: '#/components/schemas/UserRole'
        address:
          type: object
          description: 住所情報
          properties:
            postalCode:
              type: string
              description: 郵便番号（ハイフンなし）
              pattern: '^[0-9]{7}$'
              example: "1234567"
            prefecture:
              type: string
              description: 都道府県
              maxLength: 10
              example: "東京都"
            city:
              type: string
              description: 市区町村
              maxLength: 50
              example: "渋谷区"
            addressLine:
              type: string
              description: 番地・建物名
              maxLength: 100
              example: "渋谷2-2-2 渋谷タワー201"
        birthDate:
          type: string
          description: 生年月日
          format: date
          example: "1990-04-15"
        department:
          type: string
          description: 所属部署
          maxLength: 50
          example: "営業部"

    UserResponse:
      type: object
      description: ユーザー情報のレスポンス
      properties:
        id:
          type: integer
          format: int64
          description: ユーザーの一意識別子
          example: 1
        name:
          type: string
          description: ユーザーの氏名
          example: "田中太郎"
        email:
          type: string
          description: ユーザーのメールアドレス
          format: email
          example: "tanaka@example.com"
        phone:
          type: string
          description: 電話番号
          example: "09012345678"
        status:
          $ref: '#/components/schemas/UserStatus'
        role:
          $ref: '#/components/schemas/UserRole'
        address:
          type: object
          description: 住所情報
          properties:
            postalCode:
              type: string
              description: 郵便番号
              example: "1234567"
            prefecture:
              type: string
              description: 都道府県
              example: "東京都"
            city:
              type: string
              description: 市区町村
              example: "新宿区"
            addressLine:
              type: string
              description: 番地・建物名
              example: "新宿1-1-1 新宿ビル101"
        birthDate:
          type: string
          description: 生年月日
          format: date
          example: "1990-04-15"
        department:
          type: string
          description: 所属部署
          example: "開発部"
        createdAt:
          type: string
          description: アカウント作成日時
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          description: 最終更新日時
          format: date-time
          example: "2024-01-20T14:45:00Z"
        lastLoginAt:
          type: string
          description: 最終ログイン日時
          format: date-time
          example: "2024-01-25T09:15:00Z"

    UserPageResponse:
      type: object
      description: ページネーション対応のユーザー一覧レスポンス
      properties:
        content:
          type: array
          description: ユーザーデータの配列
          items:
            $ref: '#/components/schemas/UserResponse'
        pageable:
          type: object
          description: ページング情報
          properties:
            pageNumber:
              type: integer
              description: 現在のページ番号（0から開始）
              example: 0
            pageSize:
              type: integer
              description: 1ページあたりの件数
              example: 20
            sort:
              type: object
              description: ソート情報
              properties:
                sorted:
                  type: boolean
                  description: ソートされているかどうか
                  example: true
                unsorted:
                  type: boolean
                  description: ソートされていないかどうか
                  example: false
                empty:
                  type: boolean
                  description: ソート条件が空かどうか
                  example: false
            offset:
              type: integer
              description: オフセット値
              example: 0
            paged:
              type: boolean
              description: ページングが有効かどうか
              example: true
            unpaged:
              type: boolean
              description: ページングが無効かどうか
              example: false
        totalElements:
          type: integer
          description: 全体の要素数
          example: 150
        totalPages:
          type: integer
          description: 全体のページ数
          example: 8
        last:
          type: boolean
          description: 最後のページかどうか
          example: false
        first:
          type: boolean
          description: 最初のページかどうか
          example: true
        numberOfElements:
          type: integer
          description: 現在のページの要素数
          example: 20
        size:
          type: integer
          description: ページサイズ
          example: 20
        number:
          type: integer
          description: 現在のページ番号
          example: 0
        sort:
          type: object
          description: ソート情報
          properties:
            sorted:
              type: boolean
              example: true
            unsorted:
              type: boolean
              example: false
            empty:
              type: boolean
              example: false
        empty:
          type: boolean
          description: ページが空かどうか
          example: false

    ValidationErrorResponse:
      type: object
      description: バリデーションエラーの詳細情報
      properties:
        error:
          type: string
          description: エラーメッセージ
          example: "Validation failed"
        timestamp:
          type: string
          format: date-time
          description: エラー発生日時
        path:
          type: string
          description: エラーが発生したリクエストパス
          example: "/api/users"
        status:
          type: integer
          description: HTTPステータスコード
          example: 400
        validationErrors:
          type: array
          description: フィールド別のバリデーションエラー詳細
          items:
            type: object
            properties:
              field:
                type: string
                description: エラーが発生したフィールド名
                example: "email"
              rejectedValue:
                type: string
                description: 拒否された値
                example: "invalid-email"
              message:
                type: string
                description: エラーメッセージ
                example: "正しいメールアドレス形式で入力してください"